Merge Lab1 branch to master - 1st
* Basic skeleton of classes

Added the Color, Pixel, Image, and Image Editor classes.  Gave them the main functionality they will need.

* IO Helper Classes and other class changes

Created the FileWriter and FileReader class to help with file IO and creation/interpretation of Images.  
 Added a couple more skeleton functions that I need to implement/learn more about. Also changed how the braces are done.

* Added beginning of test files, can read from files now

Add coded and tested the reading from files. Everything is working for reading from files and making an image object.
Added the start of unit test files.

* Update gitignore

Updated the .gitignore for IDE-specific output files

* Deleted old tests, created new ones with SquareTest

Deleted the "handmade" test files and used the SquareTest plugin with IntelliJ to create the boilerplate code for the tests. They are running using JUnit v5.0.

* Added the .gitignore inside of the Lab1 directory

Added it inside just in case the IDE puts anything inside there that it shouldn't

* Removed any "out" files that IntelliJ may have created before. 

Removed the binary "out" folder that held the compiled Java classes. The changed .gitignore allowed for some ".idea" folder stuff to be allowed through.

* Added test for FileReaderHelper

Added test class and both tests are passing just fine.

* Able to read a file, invert it and write it.

File IO classes are working great. Minor changes to how an Image is created.   ImageEditor can now handle the cmd line arguments. Inverting an image works. Tested.

* Basic skeleton of classes

Added the Color, Pixel, Image, and Image Editor classes.  Gave them the main functionality they will need.

* IO Helper Classes and other class changes

Created the FileWriter and FileReader class to help with file IO and creation/interpretation of Images.  
 Added a couple more skeleton functions that I need to implement/learn more about. Also changed how the braces are done.

* Added beginning of test files, can read from files now

Add coded and tested the reading from files. Everything is working for reading from files and making an image object.
Added the start of unit test files.

* Deleted old tests, created new ones with SquareTest

Deleted the "handmade" test files and used the SquareTest plugin with IntelliJ to create the boilerplate code for the tests. They are running using JUnit v5.0.

* Added the .gitignore inside of the Lab1 directory

Added it inside just in case the IDE puts anything inside there that it shouldn't

* Removed any "out" files that IntelliJ may have created before. 

Removed the binary "out" folder that held the compiled Java classes. The changed .gitignore allowed for some ".idea" folder stuff to be allowed through.

* Added test for FileReaderHelper

Added test class and both tests are passing just fine.

* Able to read a file, invert it and write it.

File IO classes are working great. Minor changes to how an Image is created.   ImageEditor can now handle the cmd line arguments. Inverting an image works. Tested.

* Fixed iteration through Image for modifying it

In the Image.invert and Image.grayscale functions I was iterating through the width then the height in a nested for loop. Switched the order so that I iterate through the height then the width. (This the same was how the Image is created)

* Fixed iteration through Image while modifying it

In the Image.invert and Image.grayscale functions I was iterating through the width then the height in a nested for loop. Switched the order so that I iterate through the height then the width. (This the same was how the Image is created)

* Changed where output happens for writing the file

Put output in the ImageEditor call to write the new file. Before it was in FileWriterHelper.

* Changed braces to match guidelines

Put braces on the next line for the test classes.

* Tested the grayscale functionality

It worked like a charm.

* Beginnings of emboss functionality

Began the logic of the emboss functionality.  Added the description of what embossing and blurring an Image entails.

* More braces

Changed braces to match guidelines

* Minor adjustments to emboss function

Changed a couple of things to match the lab description of what embossing entails.

* Emboss functionality is working

Before I was iterating from the top-left to the bottom-right and overwriting the Image instance. I flipped that around and now emboss is working like a charm.

* Beginnings of motion blur functionality

Added the motion blur function to the Image and Pixel classes.  Also added error checking for the motion blur length a user must enter.

* Finished motion blur functionality

Finished motion blur. Need to test more and make unit tests.

* Deleted unnecessary comments

 Finalized all functionality of the Image Editor.

* More Unit tests

Created and started more unit tests.

* Image.toString() refactor

Removed an unnecessary String variable when converting the Image to a String. Also moved the function from FileWriterHelper class to the Image class to decouple the two.

* Pixel class refactor

Made the Pixel class comply with the DRY principle by calling the setRGB function where it would have been repeated.

* Created documentation

Created/updated documentation using the javadoc format

* Removed a function from a test

Removed the testImageToString function from the FileWriterHelperTest since that function is no longer part of the FWH class.