Got everything working and passed off the lab
Lab was passed off. Had to change the iteration for the motionblur function so that it would iterate to the last index of each row and include that value in the calculations (it was not doing so before). Added an if statement to take care of motionblur lengths of size 1.