Merge Lab2 into master (#3)
* Created Main function for SpellChecker

Just got the boiler plate down for the IDE

* Opps

Opps, didn't stage the files before commiting. Hehe

* Basic layout of classes and interfaces

Layout of classes based on the specifications.

* Slight changes to the classes

Added functions and properties that will be needed.

* Changed Words class to Dictionary

Changed class name to make more sense and got the add function working correctly.

* Changes to Main and SpellCorrector

Changed the try-catch in the SpellCorrector.useDictionary function. Added a try-catch to the Main function and made a smaller txt file to test with.

* Added another txt file for testing

All for testing

* Test.txt added

For testing.

* Changes to classes and flow of program

Got the add and find function for the Dictionary working. Changed privacy of some functions in SpellCorrector and edited the suggestSimilarWord function

* Update .gitignore

Now ignores the out directory from IDE

* Dictionary add and find refactor

Refactored the add and find functions by following the DRY principle. Everything still works great.

* Changes

Just small changes.

* Find and toString working

find and toString working. Working on equals and hashCode

* Changed the try-catch in Main and start of suggesting a word

Changed up the try catch in Main and  created needed functions and layout of suggesting a similar word. Also added verbosity to console output

* Refactored the Dictionary class

Refactored the Dictionary class to comply with what the professor wanted. There was to be no use of Strings in the implementation of INode. So I have fixed that and everything is still running like before.

* Changes to Dictionary

Changed the find function's declaration to match the Interface function, it now returns ITrie.INode. Also added an else to the find function to boost speed.

* Major changes to SpellCorrector

Implemented all the spell checking algorithms and everything seems to be working just fine. Will test some more.

* New test files

More test files

* Finished hashCode and equals for Dictionary

Finished all the overridden methods. Tested and they work correctly.

* Error handling in Main

Added a try-catch-finally block to handle anything that might make its way back to Main.

* Error handling and name changes to SpellChecker

Just basic error handling and changed the privacy of the Dictionary field to public so the toString method can be called.

* Removed

Removed the toString call to the SpellChecker's Dicitonary. It isn't need.

* Change to suggestSimilarWord

Changed it so that the lowerCase version of the word gets returned.